refactor(js): create a common kernel/bootstrap
