+++ common/Resources/ti.internal/bootstrap/invoker.js
